+++ packages/frontend-core/src/components/grid/stores/index.js
@@ -20,7 +20,6 @@ import * as Table from "./datasources/table"
 import * as ViewV2 from "./datasources/viewV2"
 import * as NonPlus from "./datasources/nonPlus"
 import * as Cache from "./cache"
-import * as Selection from "./selection"
 
 const DependencyOrderedStores = [
   Sort,
@@ -41,7 +40,6 @@ const DependencyOrderedStores = [
   Users,
   Menu,
   Pagination,
-  Selection,
   Clipboard,
   Config,
   Notifications,
