+++ packages/frontend-core/src/components/grid/stores/ui.js
@@ -7,7 +7,7 @@ import {
   MediumRowHeight,
   NewRowID,
 } from "../lib/constants"
-import { parseCellID } from "../lib/utils"
+import { getCellID, parseCellID } from "../lib/utils"
 
 export const createStores = context => {
   const { props } = context
@@ -22,6 +22,11 @@ export const createStores = context => {
   const keyboardBlocked = writable(false)
   const isDragging = writable(false)
   const buttonColumnWidth = writable(0)
+  const cellSelection = writable({
+    active: false,
+    sourceCellId: null,
+    targetCellId: null,
+  })
 
   return {
     focusedCellId,
@@ -35,6 +40,7 @@ export const createStores = context => {
     isDragging,
     buttonColumnWidth,
     selectedRows,
+    cellSelection,
   }
 }
 
@@ -47,6 +53,9 @@ export const deriveStores = context => {
     stickyColumn,
     width,
     selectedRows,
+    cellSelection,
+    columnLookupMap,
+    allVisibleColumns,
   } = context
 
   // Derive the current focused row ID
@@ -89,12 +98,67 @@ export const deriveStores = context => {
     return Object.keys($selectedRows).length
   })
 
+  // Derive whether or not we're actively selecting cells
+  const isSelectingCells = derived(cellSelection, $cellSelection => {
+    return $cellSelection.active
+  })
+
+  // Derive the full extent of all selected cells
+  const selectedCells = derived(
+    [cellSelection, rowLookupMap, columnLookupMap],
+    ([$cellSelection, $rowLookupMap, $columnLookupMap]) => {
+      const { sourceCellId, targetCellId } = $cellSelection
+      if (!sourceCellId || !targetCellId || sourceCellId === targetCellId) {
+        return {}
+      }
+      const $rows = get(rows)
+      const $allVisibleColumns = get(allVisibleColumns)
+
+      // Get source and target row and column indices
+      const sourceInfo = parseCellID(sourceCellId)
+      const targetInfo = parseCellID(targetCellId)
+
+      // Row indices
+      const sourceRowIndex = $rowLookupMap[sourceInfo.id]
+      const targetRowIndex = $rowLookupMap[targetInfo.id]
+      const lowerRowIndex = Math.min(sourceRowIndex, targetRowIndex)
+      const upperRowIndex = Math.max(sourceRowIndex, targetRowIndex)
+
+      // Column indices
+      const sourceColIndex = $columnLookupMap[sourceInfo.field]
+      const targetColIndex = $columnLookupMap[targetInfo.field]
+      const lowerColIndex = Math.min(sourceColIndex, targetColIndex)
+      const upperColIndex = Math.max(sourceColIndex, targetColIndex)
+
+      // Build map of all cells inside these bounds
+      let map = {}
+      let rowId, colName, cellId
+      for (let rowIdx = lowerRowIndex; rowIdx <= upperRowIndex; rowIdx++) {
+        for (let colIdx = lowerColIndex; colIdx <= upperColIndex; colIdx++) {
+          rowId = $rows[rowIdx]._id
+          colName = $allVisibleColumns[colIdx].name
+          cellId = getCellID(rowId, colName)
+          map[cellId] = { rowIdx, colIdx }
+        }
+      }
+      return map
+    }
+  )
+
+  // Derive the count of the selected cells
+  const selectedCellCount = derived(selectedCells, $selectedCells => {
+    return Object.keys($selectedCells).length
+  })
+
   return {
     focusedRowId,
     focusedRow,
     contentLines,
     compact,
     selectedRowCount,
+    isSelectingCells,
+    selectedCells,
+    selectedCellCount,
   }
 }
 
@@ -106,6 +170,8 @@ export const createActions = context => {
     rowLookupMap,
     rows,
     selectedRowCount,
+    cellSelection,
+    selectedCells,
   } = context
   // Keep the last selected index to use with bulk selection
   let lastSelectedIndex = null
@@ -114,6 +180,7 @@ export const createActions = context => {
   const blur = () => {
     focusedCellId.set(null)
     hoveredRowId.set(null)
+    clearCellSelection()
   }
 
   // Toggles whether a certain row ID is selected or not
@@ -159,6 +226,36 @@ export const createActions = context => {
     })
   }
 
+  const startCellSelection = sourceCellId => {
+    cellSelection.set({
+      active: true,
+      sourceCellId,
+      targetCellId: sourceCellId,
+    })
+  }
+
+  const updateCellSelection = targetCellId => {
+    cellSelection.update(state => ({
+      ...state,
+      targetCellId,
+    }))
+  }
+
+  const stopCellSelection = () => {
+    cellSelection.update(state => ({
+      ...state,
+      active: false,
+    }))
+  }
+
+  const clearCellSelection = () => {
+    cellSelection.set({
+      active: false,
+      sourceCellId: null,
+      targetCellId: null,
+    })
+  }
+
   return {
     ui: {
       actions: {
@@ -170,6 +267,15 @@ export const createActions = context => {
       actions: {
         toggleRow: toggleSelectedRow,
         bulkSelectRows,
+      },
+    },
+    selectedCells: {
+      ...selectedCells,
+      actions: {
+        start: startCellSelection,
+        update: updateCellSelection,
+        stop: stopCellSelection,
+        clear: clearCellSelection,
       },
     },
   }
@@ -189,8 +295,8 @@ export const initialise = context => {
     fixedRowHeight,
     selectedRowCount,
     menu,
-    cellSelection,
     selectedCellCount,
+    selectedCells,
   } = context
 
   // Ensure we clear invalid rows from state if they disappear
@@ -254,7 +360,7 @@ export const initialise = context => {
 
     // Clear cell selection when focusing a cell
     if (id && get(selectedCellCount)) {
-      cellSelection.actions.clear()
+      selectedCells.actions.clear()
     }
 
     // Close the menu if it was open
@@ -284,8 +390,15 @@ export const initialise = context => {
         focusedCellId.set(null)
       }
       if (get(selectedCellCount)) {
-        cellSelection.actions.clear()
+        selectedCells.actions.clear()
       }
+    }
+  })
+
+  // Clear selected rows when selecting cells
+  selectedCellCount.subscribe($selectedCellCount => {
+    if ($selectedCellCount && get(selectedRowCount)) {
+      selectedRows.set({})
     }
   })
 }
